+++ tests/Middleware/ResourceServerMiddlewareTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace T0mmy742\TokenAPI\Tests\Middleware;
 
+use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -12,7 +13,6 @@ use Slim\Psr7\Factory\ServerRequestFactory;
 use T0mmy742\TokenAPI\Exception\TokenApiException;
 use T0mmy742\TokenAPI\Middleware\ResourceServerMiddleware;
 use T0mmy742\TokenAPI\ResourceServer;
-use T0mmy742\TokenAPI\Tests\TestCase;
 
 use function json_encode;
 
@@ -22,15 +22,16 @@ class ResourceServerMiddlewareTest extends TestCase
     {
         $serverRequest = (new ServerRequestFactory())->createServerRequest('GET', '/test');
 
-        $resourceServerProphecy = $this->prophesize(ResourceServer::class);
-        $resourceServerProphecy->validateAuthenticatedRequest($serverRequest)
+        $resourceServer = $this->createMock(ResourceServer::class);
+        $resourceServer
+            ->expects($this->once())
+            ->method('validateAuthenticatedRequest')
+            ->with($serverRequest)
             ->willReturn(
                 $serverRequest
                     ->withAttribute('access_token_id', 'token_id')
                     ->withAttribute('user_id', '1')
-            )
-            ->shouldBeCalledOnce();
-        $resourceServer = $resourceServerProphecy->reveal();
+            );
 
         $resourceServerMiddleware = new ResourceServerMiddleware($resourceServer, new ResponseFactory());
 
@@ -51,11 +52,12 @@ class ResourceServerMiddlewareTest extends TestCase
     {
         $serverRequest = (new ServerRequestFactory())->createServerRequest('GET', '/badTest');
 
-        $resourceServerProphecy = $this->prophesize(ResourceServer::class);
-        $resourceServerProphecy->validateAuthenticatedRequest($serverRequest)
-            ->willThrow(new TokenApiException('Invalid token'))
-            ->shouldBeCalledOnce();
-        $resourceServer = $resourceServerProphecy->reveal();
+        $resourceServer = $this->createMock(ResourceServer::class);
+        $resourceServer
+            ->expects($this->once())
+            ->method('validateAuthenticatedRequest')
+            ->with($serverRequest)
+            ->willThrowException(new TokenApiException('Invalid token'));
 
         $resourceServerMiddleware = new ResourceServerMiddleware($resourceServer, new ResponseFactory());
 
@@ -69,5 +71,32 @@ class ResourceServerMiddlewareTest extends TestCase
         $response = $resourceServerMiddleware->process($serverRequest, $handler);
 
         $this->assertSame(json_encode(['error' => 'Invalid token']), (string) $response->getBody());
+    }
+
+    public function testBadJsonEncoding(): void
+    {
+        $serverRequest = (new ServerRequestFactory())->createServerRequest('GET', '/badTest');
+
+        $resourceServer = $this->createMock(ResourceServer::class);
+        $resourceServer
+            ->expects($this->once())
+            ->method('validateAuthenticatedRequest')
+            ->with($serverRequest)
+            ->willThrowException(new TokenApiException('Invalid token'));
+
+        $resourceServerMiddleware = new ResourceServerMiddleware($resourceServer, new ResponseFactory());
+
+        $handler = new class implements RequestHandlerInterface {
+            public function handle(ServerRequestInterface $request): ResponseInterface
+            {
+                return (new ResponseFactory())->createResponse();
+            }
+        };
+
+        $GLOBALS['json_encode_false'] = true;
+
+        $response = $resourceServerMiddleware->process($serverRequest, $handler);
+
+        $this->assertSame('JSON encoding failed', (string) $response->getBody());
     }
 }