+++ tests/ResourceServerTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace T0mmy742\TokenAPI\Tests;
 
+use PHPUnit\Framework\TestCase;
 use Slim\Psr7\Factory\ServerRequestFactory;
 use T0mmy742\TokenAPI\ResourceServer;
 use T0mmy742\TokenAPI\TokenValidator\TokenValidatorInterface;
@@ -14,10 +15,14 @@ class ResourceServerTest extends TestCase
     {
         $serverRequest = (new ServerRequestFactory())->createServerRequest('GET', '/test');
 
-        $tokenValidatorProphecy = $this->prophesize(TokenValidatorInterface::class);
-        $tokenValidatorProphecy->validateToken($serverRequest)->willReturn($serverRequest->withAttribute('test', 'OK'));
+        $tokenValidator = $this->createMock(TokenValidatorInterface::class);
+        $tokenValidator
+            ->expects($this->once())
+            ->method('validateToken')
+            ->with($serverRequest)
+            ->willReturn($serverRequest->withAttribute('test', 'OK'));
 
-        $resourceServer = new ResourceServer($tokenValidatorProphecy->reveal());
+        $resourceServer = new ResourceServer($tokenValidator);
 
         $resultRequest = $resourceServer->validateAuthenticatedRequest($serverRequest);