+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New phpstan rules
+
+### Changed
+- Remove `Crypt` trait and use `CryptInterface`
+
+### Delete
+- Remove Prophecy `phpspec/prophecy` usage for testing
 
 ## [1.0.0] - 2020-10-12
 ### Added